# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standard error response schema.",
        "example": {
          "details": { "job_id": "123" },
          "error": { "code": "NOT_FOUND", "message": "Resource not found" }
        },
        "properties": {
          "details": {
            "anyOf": [{ "additionalProperties": true, "type": "object" }, { "type": "null" }],
            "title": "Details"
          },
          "error": {
            "additionalProperties": { "type": "string" },
            "title": "Error",
            "type": "object"
          }
        },
        "required": ["error"],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Token": {
        "properties": {
          "access_token": { "title": "Access Token", "type": "string" },
          "refresh_token": { "title": "Refresh Token", "type": "string" },
          "token_type": { "default": "bearer", "title": "Token Type", "type": "string" }
        },
        "required": ["access_token", "refresh_token"],
        "title": "Token",
        "type": "object"
      },
      "TokenRefresh": {
        "properties": { "refresh_token": { "title": "Refresh Token", "type": "string" } },
        "required": ["refresh_token"],
        "title": "TokenRefresh",
        "type": "object"
      },
      "UserCreate": {
        "properties": {
          "email": { "format": "email", "title": "Email", "type": "string" },
          "password": { "title": "Password", "type": "string" }
        },
        "required": ["email", "password"],
        "title": "UserCreate",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "email": { "format": "email", "title": "Email", "type": "string" },
          "id": { "format": "uuid", "title": "Id", "type": "string" }
        },
        "required": ["id", "email"],
        "title": "UserResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "scheme": "bearer", "type": "http" } }
  },
  "info": {
    "contact": { "name": "Team 21", "url": "https://github.com/tomkabel/team21-vooglaadija" },
    "description": "REST API for user authentication, creating download jobs, tracking job status, and retrieving processed files. Authentication uses bearer JWT access tokens.",
    "license": { "name": "GPLv3", "url": "https://www.gnu.org/licenses/gpl-3.0.html" },
    "summary": "Asynchronous API for authenticated video download jobs.",
    "title": "Vooglaadija API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate with email and password and receive access/refresh JWT tokens.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh",
                  "token_type": "bearer"
                },
                "schema": { "$ref": "#/components/schemas/Token" }
              }
            },
            "description": "Authentication successful"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Incorrect email or password" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Invalid credentials or inactive user"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "details": {
                    "validation_errors": [
                      {
                        "field": "email",
                        "message": "value is not a valid email address",
                        "type": "value_error"
                      }
                    ]
                  },
                  "error": { "code": "VALIDATION_ERROR", "message": "Request validation failed" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Rate limit exceeded. Try again in 42 seconds."
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Authenticate user",
        "tags": ["auth"]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "description": "Return the authenticated user's profile based on bearer access token.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "email": "user@example.com",
                  "id": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            },
            "description": "Current user profile"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Get current user",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Exchange a valid refresh token for a new access token and refresh token pair.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [{ "$ref": "#/components/schemas/TokenRefresh" }, { "type": "null" }],
                "title": "Token Refresh"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.newaccess",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.newrefresh",
                  "token_type": "bearer"
                },
                "schema": { "$ref": "#/components/schemas/Token" }
              }
            },
            "description": "Token refreshed successfully"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Invalid or expired refresh token" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Invalid or expired refresh token"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "details": {
                    "validation_errors": [
                      { "field": "refresh_token", "message": "Field required", "type": "missing" }
                    ]
                  },
                  "error": { "code": "VALIDATION_ERROR", "message": "Request validation failed" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Rate limit exceeded. Try again in 42 seconds."
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Refresh access token",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Create a user account with email and password.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "email": "user@example.com",
                  "id": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            },
            "description": "User created successfully"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "RESOURCE_CONFLICT", "message": "Email already registered" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Email already registered"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "details": {
                    "validation_errors": [
                      {
                        "field": "password",
                        "message": "Password must be at least 8 characters",
                        "type": "value_error"
                      }
                    ]
                  },
                  "error": { "code": "VALIDATION_ERROR", "message": "Request validation failed" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Rate limit exceeded. Try again in 42 seconds."
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Register a new user",
        "tags": ["auth"]
      }
    },
...
  },
...
}